New Config, ES2020, XHR
- Switched to TOML-based config
- Now using XHR instead of request which has been depricated for a while
- Split up helper functions into seperate files
- Added a little more documentation to the code
- Added package.json and gitignore for node_modules
- Readme refactor (remove useless newlines)
