+++ src/tools/logging.js
@@ -1,0 +1,7 @@
+import { verbose } from "../index.js";
+export default function log(text, isVerbose = false) {
+  if ((isVerbose == true && verbose == true) || isVerbose == false) {
+    var d = new Date().toISOString();
+    console.log("[" + d + "] " + text);
+  }
+}
