+++ src/tools/middleware.js
@@ -1,0 +1,10 @@
+export default function parse(req, res, next) {
+  req.rawBody = "";
+  req.setEncoding("utf8");
+  req.on("data", function (chunk) {
+    req.rawBody += chunk;
+  });
+  req.on("end", function () {
+    next();
+  });
+}
